NFC Refactoring to prepare for using Comprehensive bufferization.
This PR does some NFC refactoring to prepare for using the
Comprehensive bufferization pass.
- Moves all bufferization interfaces related code into a separate
  folder within `Codegen`. This is also a place where all `Codegen`
  related interfaces can live. For example, the `ProcessorOpInterface`
  is moved into this folder as well.
- Define the callbacks to use for the comprehensive bufferization.

The bufferization itself is not enabled to wait for some other fixes
that need to land but landing this as an NFC change for now.
