+++ iree/compiler/Codegen/Common/ConvertToDestinationPassingStylePass.cpp
@@ -142,6 +142,16 @@ static Value getTiedResultForOperand(OpOperand &operand,
       return result;
     }
   }
+  // Check for loop yield values.
+  if (auto yieldOp = dyn_cast<scf::YieldOp>(operand.getOwner())) {
+    Operation *parentOp = yieldOp->getParentOp();
+    if (isa<scf::ForOp, scf::IfOp>(parentOp)) {
+      Value result = parentOp->getResult(operand.getOperandNumber());
+      if (plan.isEquivalent(result, operand.get())) {
+        return result;
+      }
+    }
+  }
   return nullptr;
 }
 
@@ -156,9 +166,6 @@ static IREE::Flow::DispatchTensorStoreOp walkUseToGetDispatchTensorStoreOp(
     llvm::DenseSet<Value> &processed) {
   Operation *user = nullptr;
   while (value.hasOneUse()) {
-    assert(!processed.count(value) &&
-           "unexpected traversal through already traversed value during "
-           "conversion to destination passing style");
     processed.insert(value);
     OpOperand &use = *value.use_begin();
     user = use.getOwner();
@@ -213,15 +220,13 @@ static LogicalResult modifyResultToUseStoreBuffer(
         "failed walk of uses to get to flow.dispatch.tensor.store op");
   }
 
-  // For now we need to change the usage to use destination passing style only
-  // if the start of the use-def chain is an init_tensor op. Might adapt this as
-  // we go along.
-  if (!isa<linalg::InitTensorOp>(resultValueOp)) {
-    return success();
-  }
-
   OpBuilder::InsertionGuard g(b);
-  b.setInsertionPoint(resultValueOp);
+  b.setInsertionPointToStart(storeOp->getBlock());
+  if (auto sourceDefiningOp = storeOp.target().getDefiningOp()) {
+    if (sourceDefiningOp->getBlock() == storeOp->getBlock()) {
+      b.setInsertionPointAfter(sourceDefiningOp);
+    }
+  }
   Value resultBuffer = getTensorLoadOpForTensorStoreOp(b, storeOp);
 
   // Now replay the instructions that are essentially doing type-conversion, in
@@ -269,9 +274,9 @@ static LogicalResult convertToDestinationPassingStyle(OpBuilder &b,
   }
 
   llvm::DenseSet<Value> processed;
-  auto walkResult =
-      funcOp.walk<WalkOrder::PreOrder>([&](Operation *op) -> WalkResult {
-        for (auto result : op->getResults()) {
+  auto walkResult = funcOp.walk<WalkOrder::PreOrder>(
+      [&](linalg::InitTensorOp initTensorOp) -> WalkResult {
+        for (auto result : initTensorOp->getResults()) {
           if (!result.getType().isa<RankedTensorType>()) continue;
           if (plan.isInStoreSet(result) && !processed.count(result)) {
             return modifyResultToUseStoreBuffer(b, result, plan, processed);
