+++ iree/compiler/Codegen/Interfaces/CMakeLists.txt
@@ -1,0 +1,69 @@
+################################################################################
+# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
+# iree/compiler/Codegen/Interfaces/BUILD                                       #
+#                                                                              #
+# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
+# CMake-only content.                                                          #
+#                                                                              #
+# To disable autogeneration for this file entirely, delete this header.        #
+################################################################################
+
+iree_add_all_subdirs()
+
+iree_cc_library(
+  NAME
+    Interfaces
+  HDRS
+    "Interfaces.h"
+  SRCS
+    "Interfaces.cpp"
+  DEPS
+    ::BufferizationInterfaces
+    ::ProcessorOpInterfaces
+  PUBLIC
+)
+
+iree_cc_library(
+  NAME
+    BufferizationInterfaces
+  HDRS
+    "BufferizationInterfaces.h"
+  SRCS
+    "BufferizationInterfaces.cpp"
+  DEPS
+    MLIRLinalgBufferizableOpInterfaceImpl
+    MLIRMemRef
+    iree::compiler::Dialect::Flow::IR
+    iree::compiler::Dialect::HAL::IR
+  PUBLIC
+)
+
+iree_cc_library(
+  NAME
+    ProcessorOpInterfaces
+  HDRS
+    "ProcessorOpInterfaces.h"
+  TEXTUAL_HDRS
+    "ProcessorOpInterfaces.cpp.inc"
+    "ProcessorOpInterfaces.h.inc"
+  SRCS
+    "ProcessorOpInterfaces.cpp"
+  DEPS
+    ::ProcessorOpInterfaceGen
+    MLIRGPUOps
+    MLIRIR
+    iree::compiler::Dialect::HAL::IR
+  PUBLIC
+)
+
+iree_tablegen_library(
+  NAME
+    ProcessorOpInterfaceGen
+  TD_FILE
+    "ProcessorOpInterfaces.td"
+  OUTS
+    -gen-op-interface-decls ProcessorOpInterfaces.h.inc
+    -gen-op-interface-defs ProcessorOpInterfaces.cpp.inc
+)
+
+### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
