+++ iree/compiler/Codegen/Interfaces/Interfaces.cpp
@@ -1,0 +1,21 @@
+// Copyright 2021 The IREE Authors
+//
+// Licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#include "iree/compiler/Codegen/Interfaces/Interfaces.h"
+
+#include "iree/compiler/Codegen/Interfaces/BufferizationInterfaces.h"
+#include "iree/compiler/Codegen/Interfaces/ProcessorOpInterfaces.h"
+
+namespace mlir {
+namespace iree_compiler {
+
+void registerCodegenInterfaces(DialectRegistry &registry) {
+  registerProcessorOpInterfaceExternalModels(registry);
+  registerBufferizationInterfaces(registry);
+}
+
+}  // namespace iree_compiler
+}  // namespace mlir
