+++ iree/compiler/Codegen/Utils/CMakeLists.txt
@@ -26,7 +26,7 @@ iree_cc_library(
     MLIRLinalg
     MLIRLinalgTransforms
     MLIRSupport
-    iree::compiler::Codegen::Dialect::ProcessorOpInterfaces
+    iree::compiler::Codegen::Interfaces::ProcessorOpInterfaces
     iree::compiler::Dialect::Flow::IR
     iree::compiler::Dialect::HAL::IR
   PUBLIC
