+++ iree/tools/CMakeLists.txt
@@ -217,7 +217,7 @@ if(${IREE_BUILD_COMPILER})
       iree::compiler::Bindings::Native::Transforms
       iree::compiler::Bindings::TFLite::Transforms
       iree::compiler::Codegen::Dialect::IREECodegenDialect
-      iree::compiler::Codegen::Dialect::ProcessorOpInterfaces
+      iree::compiler::Codegen::Interfaces::Interfaces
       iree::compiler::Dialect::Flow::IR
       iree::compiler::Dialect::Flow::Transforms
       iree::compiler::Dialect::HAL::IR::HALDialect
